requirements.txt
# SOLO-IDEAFLOW-OS: Requirements
crewai
python-dotenv
# Add your preferred LLM provider (e.g., openai, anthropic, groq)
openai
